create util and save piece as relative points.
